cloud_functions/image_generator/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
